requirements.txt
coverage==7.6.12
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest>=7.0.0
pytest-cov>=4.0.0
pyyaml>=6.0
jsonschema==4.21.1
requests>=2.25.0
